backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "bcrypt>=4.3.0",
    "docx2txt>=0.9",
    "fastapi[all]>=0.116.1",
    "greenlet>=3.2.3",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pypdf>=5.9.0",
]
